+++ examples/http-multi/index.ts
@@ -1,34 +1,26 @@
 import * as azure from "@pulumi/azure";
 
-const resourceGroup = new azure.core.ResourceGroup("example", { location: "West US 2" });
+const resourceGroup = new azure.core.ResourceGroup("example", { location: azure.Locations.WestUS2 });
 
-function makeHttpFunction(name: string, callback: azure.appservice.Callback<azure.appservice.Context<azure.appservice.HttpResponse>, azure.appservice.HttpRequest, azure.appservice.HttpResponse>) {
-    const bindings = [{
-        authLevel: "anonymous",
-        type: "httpTrigger",
-        direction: "in",
-        name: "req",
-    }, {
-        type: "http",
-        direction: "out",
-        name: "$return",
-    }];
+// Define 3 HTTP Functions which only differ by number and the hello message
+const http = [1, 2, 3].map(i =>
+    new azure.appservice.HttpFunction(`F${i}`, {
+        callback: async (context, request) => ({ status: 200, body: `Hi from F${i}` }),
+    }),
+);
 
-    return <azure.appservice.Function>{
-        name,
-        bindings,
-        callback: { callback },
-    };
-}
+// Define a timer function which will trigger every minute to keep other Function from being disposed on idle
+const warmer = new azure.appservice.TimerFunction("Warmer", {
+    schedule: { second: 0 },
+    callback: async (context, timer) => {
+        // Do nothing, it's just a warmer
+    },
+});
 
 // Create a Function App containing multiple functions
 const app = new azure.appservice.MultiCallbackFunctionApp("http-multi", {
-    resourceGroup,
-    functions: [
-        makeHttpFunction("F1", async (c, r) => ({ status: 200, body: "Hi from F1" })),
-        makeHttpFunction("F2", async (c, r) => ({ status: 200, body: "Hi from F2" }))
-    ]
+    resourceGroupName: resourceGroup.name,
+    functions: [ ...http, warmer],
 });
 
-export const url1 = app.endpoint.apply(ep => `${ep}F1`);
-export const url2 = app.endpoint.apply(ep => `${ep}F2`);
+export const urls = http.map(f => app.endpoint.apply(ep => `${ep}${f.name}`));