+++ sdk/nodejs/storage/zMixins.ts
@@ -269,7 +269,7 @@ interface QueueBindingDefinition extends appservice.BindingDefinition {
     /**
      * The storage connection string for the storage account containing the queue.
      */
-    connection: string;
+    connection: pulumi.Input<string>;
 }
 
 /**
@@ -330,16 +330,29 @@ export interface QueueHostSettings extends appservice.HostSettings {
  */
 export type QueueCallback = appservice.Callback<QueueContext, Buffer, void>;
 
+export interface QueueFunctionArgs extends appservice.CallbackArgs<QueueContext, Buffer, void> {
+    /**
+     * Defines the queue to trigger the function.
+     */
+    queue: Queue;
+};
+
 export interface QueueEventSubscriptionArgs extends appservice.CallbackFunctionAppArgs<QueueContext, Buffer, void> {
+    /**
+     * The resource group in which to create the event subscription.  If not supplied, the
+     * Queue's resource group will be used.
+     */
+    resourceGroup?: core.ResourceGroup;
+
     /**
      * The name of the resource group in which to create the event subscription. [resourceGroup] takes precedence over [resourceGroupName].
      * If none of the two is supplied, the resource group of the Storage Account will be used.
      */
     resourceGroupName?: pulumi.Input<string>;
 
-    /** 
-     * Host settings specific to the Storage Queue plugin. These values can be provided here, or defaults will 
-     * be used in their place. 
+    /**
+     * Host settings specific to the Storage Queue plugin. These values can be provided here, or defaults will
+     * be used in their place.
      */
     hostSettings?: QueueHostSettings;
 };
@@ -372,40 +385,37 @@ export class QueueEventSubscription extends appservice.EventSubscription<QueueCo
 
         const { resourceGroupName, location } = appservice.getResourceGroupNameAndLocation(args, queue.resourceGroupName);
 
-        // The queue binding does not store the storage connection string directly.  Instead, the
-        // connection string is put into the app settings (under whatever key we want). Then, the
-        // .connection property of the binding contains the *name* of that app setting key.
-        const bindingConnectionKey = "BindingConnectionAppSettingsKey";
+        super("azure:storage:QueueEventSubscription", name, new QueueFunction(name, { ...args, queue }), {
+            ...args,
+            resourceGroupName,
+            location,
+        }, opts);
 
-        const bindings: QueueBindingDefinition[] = [{
+        this.registerOutputs();
+    }
+}
+
+/**
+ * Azure Function triggered by a Storage Queue.
+ */
+export class QueueFunction extends appservice.Function<QueueContext, Buffer, void> {
+    constructor(name: string, args: QueueFunctionArgs) {
+        const bindingConnectionKey = pulumi.interpolate`${args.queue.storageAccountName}ConnectionStringKey`;
+
+        const account = pulumi.all([args.queue.resourceGroupName, args.queue.storageAccountName])
+            .apply(([resourceGroupName, storageAccountName]) =>
+                storage.getAccount({ resourceGroupName, name: storageAccountName }));
+
+        const appSettings = pulumi.all([account.primaryConnectionString, bindingConnectionKey]).apply(
+            ([connectionString, key]) => ({ [key]: connectionString }));
+
+        super(name, [<QueueBindingDefinition>{
             name: "queue",
             type: "queueTrigger",
             direction: "in",
             dataType: "binary",
-            queueName: queue.name,
+            queueName: args.queue.name,
             connection: bindingConnectionKey,
-        }];
-
-        // Place the mapping from the well known key name to the storage account connection string in
-        // the 'app settings' object.
-        const appSettingsOutput = args.appSettings || pulumi.output({});
-
-        // Place the mapping from the well known key name to the storage account connection string in
-        // the 'app settings' object.
-        const account = pulumi.all([queue.resourceGroupName, queue.storageAccountName])
-                                .apply(([resourceGroupName, storageAccountName]) =>
-                                storage.getAccount({ resourceGroupName, name: storageAccountName }));
-
-        const appSettings = pulumi.all([args.appSettings, account.primaryConnectionString]).apply(
-            ([appSettings, connectionString]) => ({ ...appSettings, [bindingConnectionKey]: connectionString }));
-
-        super("azure:storage:QueueEventSubscription", name, bindings, {
-            ...args,
-            resourceGroupName,
-            location,
-            appSettings,
-        }, opts);
-
-        this.registerOutputs();
+        }], args, appSettings);
     }
 }